requirements.txt
Flask
huggingface_hub
python-dotenv
pypdf
requests
Pillow